% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birmingham_totem_pedestal.R
\name{birmingham_totem_pedestal}
\alias{birmingham_totem_pedestal}
\title{Builds a totem pedestral sign in the style of Birmingham mobility hubs}
\usage{
birmingham_totem_pedestal(
  file_to_save_to,
  hub_name_text,
  icons_to_include_in_header = list(system.file("extdata/bus_stop.svg", package =
    "mobilityHubTools"), system.file("extdata/bicycle_repair_station.svg", package =
    "mobilityHubTools"), system.file("extdata/e-scooter-svgrepo-com.svg", package =
    "mobilityHubTools"), system.file("extdata/bicycle-electric-2.svg", package =
    "mobilityHubTools"), system.file("extdata/toilets.svg", package =
    "mobilityHubTools")),
  directions_image = system.file("extdata/birmingham_directions.png", package =
    "mobilityHubTools"),
  map_local = bristol_local_map(),
  points_of_interest_image = system.file("extdata/birmingham_points_of_interest.png",
    package = "mobilityHubTools"),
  map_regional = bristol_regional_map(),
  acknowledgements_text =
    "Map data from OpenStreetMap, available under the Open Database License. © OpenStreetMap contributors. See openstreetmap.org/copyright \\n \\nIcons from https://github.com/gmgeo/osmic https://www.svgrepo.com/svg/450115/e-scooter"
)
}
\arguments{
\item{file_to_save_to}{a character string of the file to save the pdf to}

\item{hub_name_text}{the name of the hub}

\item{icons_to_include_in_header}{a list of 5 strings of the location of the icons to include in the header}

\item{directions_image}{a character string of the location of an image showing directions}

\item{map_local}{a ggplot object}

\item{points_of_interest_image}{a character string of the location of an image showing local points of interest}

\item{map_regional}{a ggplot object}

\item{acknowledgements_text}{a character string including acknowledgements and copyright details.}
}
\value{
nothing, but outputs a pdf to the named file
}
\description{
Builds a totem pedestral sign in the style of Birmingham mobility hubs
}
