% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristol_map_base.R
\name{bristol_map_base}
\alias{bristol_map_base}
\title{Build the base layers of maps as per the Bristol mobility hub example.}
\usage{
bristol_map_base(
  amenity = rlist::list.load(system.file("data/greensborough_amenity.rdata", package =
    "mobilityHubTools")),
  building = rlist::list.load(system.file("data/greensborough_building.rdata", package =
    "mobilityHubTools")),
  highway = rlist::list.load(system.file("data/greensborough_highway.rdata", package =
    "mobilityHubTools")),
  leisure = rlist::list.load(system.file("data/greensborough_leisure.rdata", package =
    "mobilityHubTools")),
  landuse = rlist::list.load(system.file("data/greensborough_landuse.rdata", package =
    "mobilityHubTools")),
  natural = rlist::list.load(system.file("data/greensborough_natural.rdata", package =
    "mobilityHubTools")),
  railway = rlist::list.load(system.file("data/greensborough_railway.rdata", package =
    "mobilityHubTools")),
  waterway = rlist::list.load(system.file("data/greensborough_waterway.rdata", package =
    "mobilityHubTools")),
  hub_location = rlist::list.load(system.file("data/greensborough_hub_location.rdata",
    package = "mobilityHubTools")),
  map_limits = 500,
  crs_local_metres = 27700
)
}
\arguments{
\item{amenity}{osm layers as sf, with crs set to local metres.}

\item{building}{osm layers as sf, with crs set to local metres.}

\item{highway}{osm layers as sf, with crs set to local metres.}

\item{leisure}{osm layers as sf, with crs set to local metres.}

\item{landuse}{osm layers as sf, with crs set to local metres.}

\item{natural}{osm layers as sf, with crs set to local metres.}

\item{railway}{osm layers as sf, with crs set to local metres.}

\item{waterway}{osm layers as sf, with crs set to local metres.}

\item{hub_location}{sf point with name of mobility hub and location, with crs set to local metres.}

\item{map_limits}{a numeric value setting the extents of the map}

\item{crs_local_metres}{the crs to use to calculate distances in metres}
}
\value{
a map in the style of the Bristol mobility hubs
}
\description{
Build the base layers of maps as per the Bristol mobility hub example.
}
\examples{
bristol_map_base(amenity = rlist::list.load(system.file("data/greensborough_amenity.rdata", package = "mobilityHubTools")),
building = rlist::list.load(system.file("data/greensborough_building.rdata", package = "mobilityHubTools")),
highway = rlist::list.load(system.file("data/greensborough_highway.rdata", package = "mobilityHubTools")),
leisure = rlist::list.load(system.file("data/greensborough_leisure.rdata", package = "mobilityHubTools")),
landuse = rlist::list.load(system.file("data/greensborough_landuse.rdata", package = "mobilityHubTools")),
natural = rlist::list.load(system.file("data/greensborough_natural.rdata", package = "mobilityHubTools")),
railway = rlist::list.load(system.file("data/greensborough_railway.rdata", package = "mobilityHubTools")),
waterway = rlist::list.load(system.file("data/greensborough_waterway.rdata", package = "mobilityHubTools")),
hub_location = rlist::list.load(system.file("data/greensborough_hub_location.rdata", package = "mobilityHubTools")),
map_limits = 500,
crs_local_metres = 27700)
}
