% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilityHubTools-package.R
\docType{package}
\name{mobilityHubTools-package}
\alias{mobilityHubTools}
\alias{mobilityHubTools-package}
\title{mobilityHubTools: Tools for making mobility hubs}
\description{
Tools for developing maps, information displays and other materials for mobility hubs.
}
\author{
\strong{Maintainer}: James Reynolds \email{james.reynolds@atui.ie}

Authors:
\itemize{
  \item James Reynolds \email{james.reynolds@atui.ie}
}

}
\keyword{internal}
