% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis_hub_elements_map.R
\name{synthesis_hub_elements_map}
\alias{synthesis_hub_elements_map}
\title{Build a map of mobility hub elements from OSM data.}
\usage{
synthesis_hub_elements_map(
  amenity = rlist::list.load(system.file("data/greensborough_amenity.rdata", package =
    "mobilityHubTools")),
  building = rlist::list.load(system.file("data/greensborough_building.rdata", package =
    "mobilityHubTools")),
  highway = rlist::list.load(system.file("data/greensborough_highway.rdata", package =
    "mobilityHubTools")),
  leisure = rlist::list.load(system.file("data/greensborough_leisure.rdata", package =
    "mobilityHubTools")),
  landuse = rlist::list.load(system.file("data/greensborough_landuse.rdata", package =
    "mobilityHubTools")),
  natural = rlist::list.load(system.file("data/greensborough_natural.rdata", package =
    "mobilityHubTools")),
  waterway = rlist::list.load(system.file("data/greensborough_waterway.rdata", package =
    "mobilityHubTools")),
  railway = rlist::list.load(system.file("data/greensborough_railway.rdata", package =
    "mobilityHubTools")),
  hub_location = rlist::list.load(system.file("data/test_hub_location.rdata", package =
    "mobilityHubTools"))[[1]],
  map_limits = 100,
  crs_local_metres = 27700,
  hub_element_locations = tibble::tibble(lat = c(51.490910203377, 51.4907256694068,
    51.4902862077984, 51.489975, 51.4901229823077, 51.4899590085034, 51.4901216116316,
    51.4901809072928, 51.4901592856044), lon = c(-2.5638763617436, -2.56251671012625,
    -2.56227857847616, -2.563095, -2.56291934432671, -2.5629305263146, -2.56280416503089,
    -2.56231151667631, -2.56247318758945), name = c("Gainsborough Sq. Stop A",
    "Gainsborough Sq. Stop B", "Cameron Wk. stop", NA, NA, NA, NA, NA, NA), icon_filename
    = c(system.file("extdata/bus_stop.svg", 
     package = "mobilityHubTools"),
    system.file("extdata/bus_stop.svg", package = "mobilityHubTools"),
    system.file("extdata/bus_stop.svg", package = "mobilityHubTools"),
    system.file("extdata/bicycle_parking.svg", package = "mobilityHubTools"),
    system.file("extdata/bicycle_repair_station.svg", package = "mobilityHubTools"),
    system.file("extdata/e-scooter-svgrepo-com.svg", package = "mobilityHubTools"),
    system.file("extdata/bicycle-electric-2.svg", package = "mobilityHubTools"),
    system.file("extdata/toilets.svg", 
     package = "mobilityHubTools"),
    system.file("extdata/RWBA_Behinderten-WC.svg", package = "mobilityHubTools")))
)
}
\arguments{
\item{building}{osm layers as sf.}

\item{highway}{osm layers as sf.}

\item{leisure}{osm layers as sf.}

\item{landuse}{osm layers as sf.}

\item{natural}{osm layers as sf.}

\item{waterway}{osm layers as sf.}

\item{railway}{osm layers as sf.}

\item{hub_location}{sf point with name of mobility hub and location.}

\item{map_limits}{a numeric value setting the extents of the map as a buffer distance}

\item{crs_local_metres}{the numeric value representing the crs to be used for distance calculations (assumes metres)}

\item{hub_element_locations}{a tibble including values for the lat, lon, name and icon_filename to use to show hub elements}
}
\value{
a ggplot object
}
\description{
Build a map of mobility hub elements from OSM data.
}
\examples{
synthesis_hub_elements_map(
  building = rlist::list.load(system.file("data/greensborough_building.rdata", package = "mobilityHubTools")),
  highway = rlist::list.load(system.file("data/greensborough_highway.rdata", package = "mobilityHubTools")),
  leisure = rlist::list.load(system.file("data/greensborough_leisure.rdata", package = "mobilityHubTools")),
  landuse = rlist::list.load(system.file("data/greensborough_landuse.rdata", package = "mobilityHubTools")),
  natural = rlist::list.load(system.file("data/greensborough_natural.rdata", package = "mobilityHubTools")),
  waterway = rlist::list.load(system.file("data/greensborough_waterway.rdata", package = "mobilityHubTools")),
  railway = rlist::list.load(system.file("data/greensborough_railway.rdata", package = "mobilityHubTools")),
  hub_location = rlist::list.load(system.file("data/test_hub_location.rdata", package = "mobilityHubTools"))[[1]],
  map_limits = 100,
  crs_local_metres = 27700,
  hub_element_locations = tibble::tibble(lat = 51.49091020337705,
                                         lon = -2.5638763617435965,
                                         name = "Gainsborough Sq. Stop A",
                                         icon_filename = system.file(
                     "extdata/bus_stop.svg", package = "mobilityHubTools")))
}
