% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis_hub_surrounds_maps.R
\name{synthesis_hub_surrounds_map}
\alias{synthesis_hub_surrounds_map}
\title{Create a map showing the local area surrounding a mobility hub}
\usage{
synthesis_hub_surrounds_map(
  building = rlist::list.load(system.file("data/greensborough_building.rdata", package =
    "mobilityHubTools")),
  map_limits = 100,
  crs_local_metres = 27700,
  annotation_map_zoom = 18,
  annotation_map_type = "osm",
  highlight_building = c("North Bristol Advice Centre", "The Hub"),
  hub_location = rlist::list.load(system.file("data/test_hub_location.rdata", package =
    "mobilityHubTools"))[[1]]
)
}
\arguments{
\item{building}{osm layers as sf.}

\item{map_limits}{a numeric value setting the extents of the map as a buffer distance}

\item{crs_local_metres}{the numeric value representing the crs to be used for distance calculations (assumes metres)}

\item{annotation_map_zoom}{the zoom level for osm map tiles}

\item{annotation_map_type}{the osm map tile type to use}

\item{highlight_building}{character values of the names of buildings to be highlighted and labelled on the map}

\item{hub_location}{sf point with name of mobility hub and location.}
}
\value{
a ggplot object
}
\description{
Create a map showing the local area surrounding a mobility hub
}
\examples{
synthesis_hub_surrounds_map <- function(
building = rlist::list.load(system.file(
"data/greensborough_building.rdata", package = "mobilityHubTools")),
map_limits = 100,
crs_local_metres = 27700,
annotation_map_zoom = 18,
annotation_map_type = "osm",
highlight_building = c("North Bristol Advice Centre", "The Hub"),
hub_location = rlist::list.load(system.file("data/test_hub_location.rdata", package = "mobilityHubTools"))[[1]])
}
