% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis_regional_map.R
\name{synthesis_regional_map}
\alias{synthesis_regional_map}
\title{Create a map showing the regional area surrounding a mobility hub}
\usage{
synthesis_regional_map(
  amenity = rlist::list.load(system.file("data/greensborough_amenity.rdata", package =
    "mobilityHubTools")),
  building = rlist::list.load(system.file("data/greensborough_building.rdata", package =
    "mobilityHubTools")),
  highway = rlist::list.load(system.file("data/greensborough_highway.rdata", package =
    "mobilityHubTools")),
  leisure = rlist::list.load(system.file("data/greensborough_leisure.rdata", package =
    "mobilityHubTools")),
  landuse = rlist::list.load(system.file("data/greensborough_landuse.rdata", package =
    "mobilityHubTools")),
  natural = rlist::list.load(system.file("data/greensborough_natural.rdata", package =
    "mobilityHubTools")),
  waterway = rlist::list.load(system.file("data/greensborough_waterway.rdata", package =
    "mobilityHubTools")),
  railway = rlist::list.load(system.file("data/greensborough_railway.rdata", package =
    "mobilityHubTools")),
  hub_location = rlist::list.load(system.file("data/test_hub_location.rdata", package =
    "mobilityHubTools"))[[1]],
  map_limits = 1500,
  crs_local_metres = 27700,
  annotation_map_tile_type = "osm",
  annotation_map_zoom = 14,
  highlight_building_regional = c("Horfield Library", "Horfield Health Centre"),
  highlight_amenity_regional = c("University of the West of England (Frenchay Campus)",
    "St James Church"),
  highlight_leisure_regional = c("Bristol County Ground"),
  highlight_landuse_regional = c("Bonnington Walk Playing Fields")
)
}
\arguments{
\item{amenity}{osm layers as sf.}

\item{building}{osm layers as sf.}

\item{highway}{osm layers as sf.}

\item{leisure}{osm layers as sf.}

\item{landuse}{osm layers as sf.}

\item{natural}{osm layers as sf.}

\item{waterway}{osm layers as sf.}

\item{railway}{osm layers as sf.}

\item{hub_location}{sf point with name of mobility hub and location.}

\item{map_limits}{a numeric value setting the extents of the map as a buffer distance}

\item{crs_local_metres}{the numeric value representing the crs to be used for distance calculations (assumes metres)}

\item{annotation_map_tile_type}{the osm map tile type to use. If NA a background map is built from the osm sf layers}

\item{annotation_map_zoom}{the zoom level for osm map tiles}

\item{highlight_building_regional}{character values of the names of buildings to be highlighted and labelled on the map}

\item{highlight_amenity_regional}{character values of the names of amenities to be highlighted and labelled on the map}

\item{highlight_leisure_regional}{character values of the names of leisure facilities to be highlighted and labelled on the map}

\item{highlight_landuse_regional}{character values of the names of landuses to be highlighted and labelled on the map}
}
\value{
a ggplot object
}
\description{
Create a map showing the regional area surrounding a mobility hub
}
\examples{
synthesis_regional_map(amenity = rlist::list.load(system.file(
"data/greensborough_amenity.rdata", package = "mobilityHubTools")),
building = rlist::list.load(system.file(
"data/greensborough_building.rdata", package = "mobilityHubTools")),
highway = rlist::list.load(system.file(
"data/greensborough_highway.rdata", package = "mobilityHubTools")),
leisure = rlist::list.load(system.file(
"data/greensborough_leisure.rdata", package = "mobilityHubTools")),
landuse = rlist::list.load(system.file(
"data/greensborough_landuse.rdata", package = "mobilityHubTools")),
natural = rlist::list.load(system.file(
"data/greensborough_natural.rdata", package = "mobilityHubTools")),
waterway = rlist::list.load(system.file(
"data/greensborough_waterway.rdata", package = "mobilityHubTools")),
railway = rlist::list.load(system.file(
"data/greensborough_railway.rdata", package = "mobilityHubTools")),
hub_location = rlist::list.load(system.file(
"data/test_hub_location.rdata", package = "mobilityHubTools"))[[1]],
map_limits = 1500,
crs_local_metres = 27700,
annotation_map_tile_type = "osm",
annotation_map_zoom = 14,
highlight_building_regional = c("Horfield Library", "Horfield Health Centre"),
highlight_amenity_regional =
c("University of the West of England (Frenchay Campus)", "St James Church"),
highlight_leisure_regional = c("Bristol County Ground"),
highlight_landuse_regional = c("Bonnington Walk Playing Fields"))
}
