% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesis_totem.R
\name{synthesis_totem}
\alias{synthesis_totem}
\title{Builds a totem pedestal that is a sythesis version of various styles}
\usage{
synthesis_totem(
  file_to_save_to,
  element1_what_is_this_top_text = "Mobility Hub / Local Travel Point",
  element2_where_is_it_top_text,
  element2_where_is_it_top_text_detail,
  element3_what_is_here_top = list(system.file("extdata/bus_stop.svg", package =
    "mobilityHubTools"), system.file("extdata/bicycle_parking.svg", package =
    "mobilityHubTools"), system.file("extdata/bicycle_repair_station.svg", package =
    "mobilityHubTools"), system.file("extdata/e-scooter-svgrepo-com.svg", package =
    "mobilityHubTools"), system.file("extdata/bicycle-electric-2.svg", package =
    "mobilityHubTools"), system.file("extdata/toilets.svg", package =
    "mobilityHubTools"), system.file("extdata/RWBA_Behinderten-WC.svg", package =
    "mobilityHubTools"), 
     system.file("extdata/No_image.svg", package =
    "mobilityHubTools"), system.file("extdata/No_image.svg", package =
    "mobilityHubTools"), system.file("extdata/No_image.svg", package =
    "mobilityHubTools")),
  element4_up_what_is_nearby,
  element4_left_what_is_nearby,
  element4_right_what_is_nearby,
  element4_layout_matrix = rbind(c(1, 1, 1, 1, 1), c(2, 2, 2, 2, 2), c(3, 3, 3, 3, 3)),
  element5_regional_map = synthesis_regional_map(),
  element6_what_is_this_bottom_text =
    "What is this: \\nWelcome to this Mobility Hub / Local Travel Point,\\n a location where shared mobility and other services\\n offer a range of options to support sustainable travel. \\n\\nWhat is here: (see left)\\nThis mobility hub includes...",
  element7_hub_elements_map = synthesis_hub_elements_map(),
  element8_how_can_you_travel =
    "How can you travel from here: (see maps below) \\n By transit ....\\n On foot...\\n By bike....(see map below right).",
  element9_hub_surrounds_map = synthesis_hub_surrounds_map(),
  element10_what_is_nearby_bottom_text =
    "What is nearby: (see maps above)\\n\\n\\n\\n \\nWhere am I:\\n ...\\nThe latitude is ... and the longitude is ....\\nwhat3words.com calls this location ...\\n\\n ",
  element11_transport_map_small = synthesis_hub_surrounds_map(map_limits = 100,
    annotation_map_zoom = 17, annotation_map_type = "osmtransport", highlight_building =
    ""),
  element12_transport_map_large = synthesis_hub_surrounds_map(map_limits = 2000,
    annotation_map_zoom = 13, annotation_map_type = "osmtransport", highlight_building =
    ""),
  element13_cycle_map = synthesis_hub_surrounds_map(map_limits = 2000,
    annotation_map_zoom = 13, annotation_map_type = "opencycle", highlight_building = ""),
  element14_acknowledgements_text =
    "Map data from OpenStreetMap, available under the Open Database License. \\n © OpenStreetMap contributors. See openstreetmap.org/copyright \\n \\n Map tiles from: OSM standard layer © OpenStreetMap contributors.; \\n and Thunderforest Open CycleMap and Transport layers by Andy Allan, https://www.thunderforest.com/ \\n \\nIcons from https://github.com/gmgeo/osmic https://www.svgrepo.com/svg/450115/e-scooter \\n and https://commons.wikimedia.org/wiki/File:RWBA_Behinderten-WC.svg"
)
}
\arguments{
\item{file_to_save_to}{a character string of the file to save the png to}

\item{element1_what_is_this_top_text}{a character string displayed at the top of the totem}

\item{element2_where_is_it_top_text}{a character string of the name of the hub location}

\item{element2_where_is_it_top_text_detail}{a character string describing the location of the hub (address details)}

\item{element3_what_is_here_top}{a list of 10 strings of the location of the icons to include in the header, including blanks}

\item{element4_up_what_is_nearby}{a character string (with arrows) to indicate what is nearby in front of the totem}

\item{element4_left_what_is_nearby}{a character string (with arrows) to indicate what is nearby to the left of the totem}

\item{element4_right_what_is_nearby}{a character string (with arrows) to indicate what is nearby to the right of the totem}

\item{element4_layout_matrix}{a matrix of characters setting the layout of the up, left and right directions. Defaults to 3 lines}

\item{element5_regional_map}{a ggplot object}

\item{element6_what_is_this_bottom_text}{a character string describing what the hub is and what is nearby}

\item{element7_hub_elements_map}{a ggplot object}

\item{element8_how_can_you_travel}{a character string describing travel options from the hub.}

\item{element9_hub_surrounds_map}{a ggplot object}

\item{element10_what_is_nearby_bottom_text}{a character string describing what is nearby and other relevant details.}

\item{element11_transport_map_small}{a ggplot object}

\item{element12_transport_map_large}{a ggplot object}

\item{element13_cycle_map}{a ggplot object}

\item{element14_acknowledgements_text}{a character string including acknowledgements and copyright details.}
}
\value{
nothing, but outputs a pdf to the named file
}
\description{
Builds a totem pedestal that is a sythesis version of various styles
}
