% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristol_map_local.R
\name{bristol_local_map}
\alias{bristol_local_map}
\title{Build the larger scale local map as per the Bristol mobility hub example.}
\usage{
bristol_local_map(
  amenity,
  building,
  highway,
  leisure,
  landuse,
  natural,
  railway,
  waterway,
  hub_location,
  map_limits = 500,
  crs_local_metres = 27700,
  highlight_building = c("North Bristol Advice Centre", "The Hub"),
  highlight_amenity = c("Stoke Park Primary School",
    "Saint Mary Magdalen and Saint Francis Lockleaze", "St James Church"),
  highlight_leisure = c("Lockleaze Youth And Play Space", "Gainsborough Square")
)
}
\arguments{
\item{amenity}{osm layers as sf, with crs set to local metres.}

\item{building}{osm layers as sf, with crs set to local metres.}

\item{highway}{osm layers as sf, with crs set to local metres.}

\item{leisure}{osm layers as sf, with crs set to local metres.}

\item{landuse}{osm layers as sf, with crs set to local metres.}

\item{natural}{osm layers as sf, with crs set to local metres.}

\item{railway}{osm layers as sf, with crs set to local metres.}

\item{waterway}{osm layers as sf, with crs set to local metres.}

\item{hub_location}{sf point with name of mobility hub and location, with crs set to local metres.}

\item{map_limits}{a numeric value setting the extents of the map}

\item{crs_local_metres}{the numeric value representing the crs for local metres}

\item{highlight_building}{character values of the names of buildings to be highlighted and labelled on the map}

\item{highlight_amenity}{character values of the names of amenities to be labelled on the map}

\item{highlight_leisure}{character values of the names of leisure facilities to be labelled on the map}
}
\value{
a ggplot object of a map in the style of the Bristol mobility hubs
}
\description{
Build the larger scale local map as per the Bristol mobility hub example.
}
