% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristol_map_base.R
\name{bristol_map_base}
\alias{bristol_map_base}
\title{Build the base layers of maps as per the Bristol mobility hub example.}
\usage{
bristol_map_base(
  amenity,
  building,
  highway,
  leisure,
  landuse,
  natural,
  railway,
  waterway,
  hub_location,
  map_limits = 500,
  crs_local_metres = 27700
)
}
\arguments{
\item{amenity}{osm layers as sf, with crs set to local metres.}

\item{building}{osm layers as sf, with crs set to local metres.}

\item{highway}{osm layers as sf, with crs set to local metres.}

\item{leisure}{osm layers as sf, with crs set to local metres.}

\item{landuse}{osm layers as sf, with crs set to local metres.}

\item{natural}{osm layers as sf, with crs set to local metres.}

\item{railway}{osm layers as sf, with crs set to local metres.}

\item{waterway}{osm layers as sf, with crs set to local metres.}

\item{hub_location}{sf point with name of mobility hub and location, with crs set to local metres.}

\item{map_limits}{a numeric value setting the extents of the map}

\item{crs_local_metres}{the crs to use to calculate distances in metres}
}
\value{
a map in the style of the Bristol mobility hubs
}
\description{
Build the base layers of maps as per the Bristol mobility hub example.
}
