% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristol_totem_pedestal.R
\name{bristol_totem_pedestal}
\alias{bristol_totem_pedestal}
\title{Builds a totem pedestral sign in the style of Bristol mobility hubs}
\usage{
bristol_totem_pedestal(
  file_to_save_to = "layout_test.pdf",
  hub_name_text = "Gainsborough Square",
  icons_to_include_in_header = list(system.file("extdata/bristol_blank.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_blank.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_blank.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_bus.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_share_car.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_charging.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_toilets.png", package =
    "mobilityHubTools"), 
     system.file("extdata/bristol_blank.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_blank.png", package =
    "mobilityHubTools"), system.file("extdata/bristol_blank.png", package =
    "mobilityHubTools")),
  directions_image = system.file("extdata/bristol_greensborough_directions.png", package
    = "mobilityHubTools"),
  map_local = map_local_test,
  facilities_image = system.file("extdata/bristol_facilities.png", package =
    "mobilityHubTools"),
  map_regional = map_regional_test,
  logo_image = system.file("extdata/bristol_blank.png", package = "mobilityHubTools")
)
}
\arguments{
\item{file_to_save_to}{a character string of the file to save the png to}

\item{hub_name_text}{the name of the hub}

\item{icons_to_include_in_header}{a list of 10 strings of the location of the icons to include in the header, including blanks}

\item{directions_image}{a character string of the location of an image showing directions}

\item{map_local}{the output of the bristol_map_local function}

\item{facilities_image}{a character string of the location of an image showing details of the hub facilities}

\item{map_regional}{the output of the bristol_map_regional function}

\item{logo_image}{a character string of the location of an image showing agency logos}
}
\value{
nothing, but outputs a png to the named file
}
\description{
Builds a totem pedestral sign in the style of Bristol mobility hubs
}
