% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_to_lonlat.R
\name{geometry_to_lonlat}
\alias{geometry_to_lonlat}
\title{Add lat and lon as numeric variables to an sf object}
\usage{
geometry_to_lonlat(x)
}
\arguments{
\item{x}{an sf object}
}
\value{
an sf object with the lat and lon included as numeric variables
}
\description{
Add lat and lon as numeric variables to an sf object
}
